backend_consolidate_20251015-1755/backend_reorganized/backend/app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.12.4
SQLAlchemy==2.0.30
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
